MELLM_pipeline/requirements.txt
absl-py==2.3.0
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albucore==0.0.16
albumentations==1.4.14
altair==5.5.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
async-lru==2.0.5
attrs==25.3.0
audioread==3.0.1
av==14.3.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
boto3==1.40.54
botocore==1.40.54
braceexpand==0.1.7
cachetools==6.2.0
cfgv==3.4.0
chumpy==0.70
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
cython==3.1.1
datasets==4.2.0
decorator==5.2.1
deepspeed==0.16.5
defusedxml==0.7.1
diffusers==0.30.0
dill==0.3.8
distlib==0.4.0
docstring-parser==0.16
easydict==1.13
einops==0.8.1
et-xmlfile==2.0.0
eval-type-backport==0.2.2
face-alignment==1.4.1
fastapi==0.115.12
fastjsonschema==2.21.2
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flask==3.1.2
flatbuffers==25.9.23
fonttools==4.57.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.45
glasses-detector==1.0.3
gradio==4.29.0
gradio-client==0.16.1
gradio-imageslider==0.0.20
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
h5py==3.14.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
identify==2.6.15
imageio==2.34.2
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
insightface==0.7.3
isoduration==20.11.0
itsdangerous==2.2.0
jieba==0.42.1
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.12.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.8
lark==1.3.0
lazy-loader==0.4
librosa==0.11.0
llamafactory==0.9.3.dev0
llvmlite==0.44.0
loguru==0.7.3
markdown==3.8
markupsafe==2.1.5
matplotlib==3.10.1
mistune==3.1.4
modelscope==1.30.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
mypy==1.18.2
mypy-extensions==1.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
notebook-shim==0.2.4
numba==0.61.2
numpy==2.2.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvitop==1.4.2
onnx==1.18.0
onnxruntime==1.23.1
openai==1.72.0
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opencv-python-headless==4.12.0.88
openface-test==0.1.26
openpyxl==3.1.5
opentsne==1.0.2
orjson==3.10.16
pandas==2.2.3
pandocfilters==1.5.1
pathspec==0.12.1
peft==0.6.2
pillow==11.3.0
plotly==6.3.1
pooch==1.8.2
pre-commit==4.3.0
prettytable==3.16.0
prometheus-client==0.23.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydeck==0.9.1
pydub==0.25.1
pyglet==2.1.6
pylance==0.38.2
pyopengl==3.1.0
pyparsing==3.2.3
pyrender==0.1.39
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytorchcv==0.0.73
pytz==2025.2
pyyaml==6.0.2
qwen-vl-utils==0.0.10
referencing==0.36.2
regex==2024.11.6
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rouge-chinese==1.0.3
rpds-py==0.26.0
ruff==0.14.1
s3fs==0.4.2
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.4.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
seamless-interaction==0.1.0
semantic-version==2.10.0
send2trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.34.1
shellingham==1.5.4
shtab==1.7.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
smplx==0.1.28
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
spiga==0.0.6
sse-starlette==2.2.1
starlette==0.46.1
streamlit==1.49.1
stringzilla==3.12.5
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.12.0
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchvision==0.22.1+cu126
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
trimesh==3.6.18
triton==3.3.1
trl==0.9.6
typer==0.15.2
types-python-dateutil==2.9.0.20251008
tyro==0.8.14
tzdata==2025.2
uri-template==1.3.0
uvicorn==0.34.0
virtualenv==20.35.3
wandb==0.21.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==11.0.3
werkzeug==3.1.3
wget==3.2
xformers==0.0.31.post1
xlsxwriter==3.2.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.19.0
zipp==3.22.0